+++ src/org/openpatch/scratch/internal/Stamp.java
@@ -9,10 +9,10 @@ import processing.core.PGraphics;
 public class Stamp {
 
   public Image image;
-  private float x;
-  private float y;
+  private double x;
+  private double y;
   private RotationStyle style;
-  private float degrees;
+  private double degrees;
 
   public Stamp(Image image, double x2, double y2) {
     this(image, 0, x2, y2, RotationStyle.DONT);
@@ -20,24 +20,24 @@ public class Stamp {
 
   public Stamp(Image image, double degrees, double x, double y, RotationStyle style) {
     this.image = image;
-    this.x = (float) x;
-    this.y = (float) y;
+    this.x = x;
+    this.y = y;
     this.style = style;
-    this.degrees = (float) degrees;
+    this.degrees = degrees;
   }
 
   public void draw(PGraphics g) {
     g.push();
     g.imageMode(PConstants.CENTER);
     g.translate(
-        this.x + Window.getInstance().getWidth() / 2,
-        -this.y + Window.getInstance().getHeight() / 2);
+        (float) (this.x + Window.getInstance().getWidth() / 2),
+        (float) (-this.y + Window.getInstance().getHeight() / 2));
     this.degrees -= 90;
     switch (this.style) {
       case DONT:
         break;
       case ALL_AROUND:
-        g.rotate(PApplet.radians(this.degrees));
+        g.rotate(PApplet.radians((float) this.degrees));
         break;
       case LEFT_RIGHT:
         if (this.degrees > -90 && this.degrees < 90) {
@@ -48,10 +48,10 @@ public class Stamp {
         break;
     }
     g.tint(
-        this.image.tint.getRed(),
-        this.image.tint.getGreen(),
-        this.image.tint.getBlue(),
-        this.image.transparency);
+        (float) this.image.tint.getRed(),
+        (float) this.image.tint.getGreen(),
+        (float) this.image.tint.getBlue(),
+        (float) this.image.transparency);
     g.image(this.image.image, 0, 0);
     g.noTint();
     g.pop();
